The error handling in post and delete added.

+++ controllers/clubs.js
@@ -34,31 +34,66 @@ export const postClub = async (req,res) => {
 export const putClub = async (req,res) => {
 
     const body = req.body;
+    var club;
 
     try {
-        await clubModel.updateOne({ _id: body._id }, req.body);
-        res.setHeader("ContentType", "application/json");
-        res.status(200).json(await clubModel.findOne(body));
-
-    } catch (error) {
-        res.setHeader("ContentType", "application/json");
-        res.status(error.status).json({ message: error.message });     
-    }
-
-};
-
-export const delClub = async (req,res) => {
-
-    const body = req.body;
-
-    try {
-        await clubModel.deleteOne(body);
-        res.setHeader("ContentType", "application/json");
-        res.status(200).json(body);
+        club = await clubModel.findOne({ _id: body._id});
         
     } catch (error) {
         res.setHeader("ContentType", "application/json");
         res.status(error.status).json({ message: error.message });    
 
+    }
+
+    if(club!=null)
+    {
+        try {
+            await clubModel.updateOne({ _id: body._id }, req.body);
+            res.setHeader("ContentType", "application/json");
+            res.status(200).json(await clubModel.findOne(body));
+
+        } catch (error) {
+            res.setHeader("ContentType", "application/json");
+            res.status(error.status).json({ message: error.message });     
+        }
+    }
+    else
+    {
+        res.setHeader("ContentType", "application/json");
+        res.status(406).json({ message: "The Club doesn't exsist."});
+    }
+};
+
+export const delClub = async (req,res) => {
+
+    const body = req.body;
+    var club;
+
+    try {
+        club = await clubModel.findOne({ _id: body._id});
+        
+    } catch (error) {
+        res.setHeader("ContentType", "application/json");
+        res.status(error.status).json({ message: error.message });    
+
+    }
+
+    if(club!=null)
+    {
+        try {
+            await clubModel.deleteOne(body);
+            res.setHeader("ContentType", "application/json");
+            res.status(200).json(body);
+        
+        } catch (error) {
+            res.setHeader("ContentType", "application/json");
+            res.status(error.status).json({ message: error.message });    
+
+        }
+    }
+    else
+    {
+        res.setHeader("ContentType", "application/json");
+        res.status(406).json({ message: "The Club doesn't exsist."});
     }
 };
